housekeeping: refactor home screen into folder
